+++ IntellijWorkspace/SkyBlock-Plugin/src/main/java/me/fan87/commonplugin/players/skill/impl/SkillFarming.java
@@ -28,7 +28,7 @@ public class SkillFarming extends SBSkill {
     }
 
     @Override
-    public int getFortuneValue(int level) {
+    public double getFortuneValue(int level) {
         return level*4;
     }
