+++ IntellijWorkspace/SkyBlock-Plugin/src/main/java/me/fan87/commonplugin/players/skill/impl/SkillForaging.java
@@ -27,7 +27,7 @@ public class SkillForaging extends SBSkill {
     }
 
     @Override
-    public int getFortuneValue(int level) {
+    public double getFortuneValue(int level) {
         return level*4;
     }
